refactor(fleet): move the fan-out core to edda-store so the pack can use it
GH-408 adds a Fleet section to the SessionStart pack and says it depends
on the fan-out core from GH-407. It cannot: the core lives in edda-cli,
which is [[bin]] with no [lib], so nothing else in the workspace can link
it. The pack is rendered in edda-bridge-claude, a lib crate.

The split is mechanism vs sentence. fan_out, group_by_project, FleetHit
and FleetMiss move to edda_store::fleet, beside the registry they already
read — no new dependencies; edda-store already had anyhow, serde_json and
tempfile. The CLI keeps print_misses, empty_summary, elsewhere_hint and
json_envelope, which are how the CLI talks, not how a fan-out works.

Pure move, no behaviour change. Test accounting is exact: the bin drops
from 216 to 211 and edda-store gains those five. Workspace 1972 passed,
0 failed, clippy 0.
